plugin/bundle: Correct bundle delay behavior
I ran into an issue when testing an earlier change:

https://github.com/open-policy-agent/opa/actions/runs/16646981900/job/47110035165

I found that this test generated around 100,000 lines of errors showing
the bundle downloader running.

This can be be tested using:

```
go test -v ./v1/plugins/bundle -count=1 2>&1 | grep -c "request failed"
```

This commit closes managers and plugins correctly, and also addresses
the underlying issue where MaxDelaySeconds was used as an ns value, and
not a seconds value (causing retrys to take place many times a second as
we see in the linked logs).

Signed-off-by: Charlie Egan <charlie@styra.com>