Avoid unnecessary use of customized transport
Unless specific options are used with http.send that necessitate a
custom http.Transport configuration, the default transport should be
used to perform the operation.

Under load, policies that use http.send were encountering errors and
high throughput variability. While troubleshooting, we noted a high
number of sockets in the TIME_WAIT state. This indicated that client
connections were not getting reused during policy evaluation. We
attempted to address this by increasing MaxIdleConns and
MaxIdleConnsPerHost on the default transport. Unfortunately, this did
not help.

When we looked through the code, we discovered that nearly every path
through http.send creates a new transport. When transports aren't
reused, connection pooling is effectively disabled. These new transports
also disable keep-alive. Either one of these effectively forces a new
client connections and associated TLS handshake for each request.

We found one existing path through the code that constructs an
http.Client without a transport: explicitly setting tls_use_system_certs
to false. When we made this change to our policy, the default transport
was used, we no longer encountered errors, and the throughput was higher
and more stable.

This change removes the block that explicitly set tlsUseSystemCerts to
true when tls_use_system_certs was unset and no other CA cert options
were used. This results in the use of the default transport when TLS
options are not set and, by default, the default transport will use the
system certificate pool.

Signed-off-by: Matthew Sykes <matthew.sykes@fmr.com>