build: bump go (1.24.6 -> 1.24.7) (#7881)
* build: bump go (1.24.6 -> 1.24.7)

https://groups.google.com/g/golang-announce/c/PtW9VW21NPs

* ci: run go stuff if .go-version changed
* Makefile: change debian base image

There is no 1.24.7-bullseye, because that distribution is EOL

* build: remove GOOS from Makefile call

Some change either in the debian trixie golang image, or in golang's
toolchain made this necessary: Prior to this commit, `GOOS` was set to
windows, causing a cross-build. We don't want a crossbuild for this
utility, we want to invoke the Linux binary (on the Linux host) to do
its job on a Windows binary (cross-built from the Linux host).

---------

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>