docs: add Compile API data filtering docs (#7939)
* docs/policy-performance: remove "n/a" in "Reason", add naked ref

Upon reading this again, I found it confusing. I think it's slightly
less so now.

Also adds a row for the lonely statements supported in the next release.

* docs: integrate data filtering docs

This is everything we've had in EOPA, with updated cross-references,
updated annotations (custom -> compile) and headers.

This is work done in another repository, so I'll collect co-authors on
this commit message's trailers.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>
Co-authored-by: Philip Conrad <philip@chariot-chaser.net>
Co-authored-by: Chris Hendrix <chendrix@users.noreply.github.com>