Ref String() and greatly improved builtin lookup cost (#7961)
I was meant to do slides for a talk, but here we are.
Fun exercise in "writing" a string without actually writing
one, but rather by trying to exclude all other possible
built-in function names to compare against ref as early as
possible in order to make the operation inexpensive. And it
worked! Down from a few allocations (worst case) and at least
3 digit ns/op, to no allocations and lower end of the 2 digit
ns/op scale.

For reference, this PR accounts for 700K allocations gone
in `regal lint bunde`, as we do millions of builtin function
calls as part of that benchmark.

Another 300K+ allocs gone by revisiting the `canSkipBctx`
field on the AST builtin struct, and making it public sp that
client like Regal may benefit. And we did :)

Signed-off-by: Anders Eknert <anders@eknert.com>