perf: Zero alloc AST store lookups of interned path terms
We should invest some time in understanding how to better utilize this in
OPA. But for the time being, this at least allows clients like Regal to avoid
allocating for lookups when terms in the path are already interned by them.

Also added the string "10" to the set of interned strings, as I found that was
missing :) And a cosmetic fix in the parser code.

Benchmarks found in the code.

Signed-off-by: Anders Eknert <anders@eknert.com>