server/failtracer: don't assume only being fed two-elem calls
This was overly specific before, and quickly failed when fail events had
not been on of the right sort. Now, we're more lenient.

These would happen, as in https://github.com/orgs/open-policy-agent/discussions/722#discussioncomment-14812737,
when a fail event happened that's unrelated to unknowns lookup, but a
valid failure nonetheless. So this will not lead to a hint. Only fail
events of the previously-expected form could yield hints.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>