Avoid unnecessary use of custom http.Transport by http.send (#7927)
* Ensure HTTP transport is setup when SNI is used

An operation specific http.Transport is not always setup in http.send.
In these cases, the ServerName populated in the TLS configuration is not
used by the client. Existing tests did not exercise this path because
they would always configure additional TLS options that resulted in an
operation specific transport.

This change moves the Host header and tls_server_name processing closer
to the rest of the TLS logic and sets isTLS to true to ensure a custom
transport is configured.

Signed-off-by: Matthew Sykes <matthew.sykes@fmr.com>

* Simplify http.Transport construction for http.send

When unix domain sockets are used, the URL scheme is forced to "http".
This change effectively disables TLS and removes the need to wire a
tls.Config instance.

Signed-off-by: Matthew Sykes <matthew.sykes@fmr.com>

* Avoid unnecessary use of customized transport

Unless specific options are used with http.send that necessitate a
custom http.Transport configuration, the default transport should be
used to perform the operation.

Under load, policies that use http.send were encountering errors and
high throughput variability. While troubleshooting, we noted a high
number of sockets in the TIME_WAIT state. This indicated that client
connections were not getting reused during policy evaluation. We
attempted to address this by increasing MaxIdleConns and
MaxIdleConnsPerHost on the default transport. Unfortunately, this did
not help.

When we looked through the code, we discovered that nearly every path
through http.send creates a new transport. When transports aren't
reused, connection pooling is effectively disabled. These new transports
also disable keep-alive. Either one of these effectively forces a new
client connections and associated TLS handshake for each request.

We found one existing path through the code that constructs an
http.Client without a transport: explicitly setting tls_use_system_certs
to false. When we made this change to our policy, the default transport
was used, we no longer encountered errors, and the throughput was higher
and more stable.

This change removes the block that explicitly set tlsUseSystemCerts to
true when tls_use_system_certs was unset and no other CA cert options
were used. This results in the use of the default transport when TLS
options are not set and, by default, the default transport will use the
system certificate pool.

Signed-off-by: Matthew Sykes <matthew.sykes@fmr.com>

---------

Signed-off-by: Matthew Sykes <matthew.sykes@fmr.com>