perf: 10% faster compilation
Or something in that ballpark. This is accomplished by tweaking
existing walk/visitor implementations where appropriate, and by
introducing a new visitor type that while more limited, is able
to perform much better where it can be used. Most impressive of
the improvements here is the new `WalkRules` implementation, as
the benchmark included in this change shows. The most impactful
change is however the improved use of the various `VarVisitor`s
used during the compilation process.

Compiling the Regal bundle:

**OPA v1.11.0 (main)**
```
67207511 ns/op  42652069 B/op    1032387 allocs/op
```

**Now**
```
62502125 ns/op  35918354 B/op     929634 allocs/op
```

Signed-off-by: Anders Eknert <anders.eknert@apple.com>