oracle: Refactor oracle to be more extensible (#8105)
* oracle: Refactor oracle to be more extensible

Oracle is going to get some updates with the upcoming string
interpolation work and I'd like to make it easier to work on with tests
etc.

I have restructured this into a 'Manager', with a number of 'Locators'
for different ast constructs. I have tried to add some basic tests for
the locators but pausing here for feedback.

I hope to continue to improve this as I add support for string interp.

Signed-off-by: Charlie Egan <charlie_egan@apple.com>

* Correctly find vars when there is shadowing

Signed-off-by: Charlie Egan <charlie_egan@apple.com>

* Remove Applicable(), duplicated in Find

This was meant to be easy to test and show what each locator was for,
but in the end is mostly just duplicated in Find().

Signed-off-by: Charlie Egan <charlie_egan@apple.com>

---------

Signed-off-by: Charlie Egan <charlie_egan@apple.com>