docs/ocp: update "concepts" for v0.2.0
The stack automounting introduced in OCP v0.2.0 made the example break.
Since it's only been a sketch in the first place, this made it harder
for people to follow.

How, a more complete version of this is a test case in OCP:
https://github.com/open-policy-agent/opa-control-plane/pull/215

In a follow-up, we'll most likely update the docs to use stack auto-
mounting, and document that.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>