benchmarks: add .nojekyll to skip jekyll build in GHA
Most external CI workflows "deploy" to GitHub Pages by committing the build output
to the gh-pages branch of the repository, and typically include a .nojekyll file.
When this happens, the GitHub Actions workflow will detect the state that the
branch does not need a build step, and will execute only the steps necessary to
deploy the site to GitHub Pages servers.

https://docs.github.com/en/pages/getting-started-with-github-pages/configuring-a-publishing-source-for-your-github-pages-site?versionId=free-pro-team%40latest&productId=pages&restPage=setting-up-a-github-pages-site-with-jekyll%2Cabout-github-pages-and-jekyll#troubleshooting-publishing-from-a-branch

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>